Refactor existing rules
