+++ service/custom-ktlint-rules/src/main/kotlin/evaka/ktlint/CustomRuleSetProvider.kt
@@ -2,13 +2,13 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-package fi.espoo.evaka.ktlint
+package evaka.ktlint
 
 import com.pinterest.ktlint.cli.ruleset.core.api.RuleSetProviderV3
 import com.pinterest.ktlint.rule.engine.core.api.RuleProvider
 import com.pinterest.ktlint.rule.engine.core.api.RuleSetId
 
-class CustomRuleSetProvider : RuleSetProviderV3(RuleSetId("custom-ktlint-rules")) {
+class CustomRuleSetProvider : RuleSetProviderV3(RuleSetId("evaka")) {
     override fun getRuleProviders(): Set<RuleProvider> =
         setOf(RuleProvider { NoJUnit4Imports() }, RuleProvider { NoPrint() })
 }
