+++ service/custom-ktlint-rules/src/main/resources/META-INF/services/com.pinterest.ktlint.cli.ruleset.core.api.RuleSetProviderV3
@@ -1,1 +1,1 @@
-fi.espoo.evaka.ktlint.CustomRuleSetProvider
+evaka.ktlint.CustomRuleSetProvider
